Split wasm consumer tests per API area
Mirror the C consumer layout: break the single wasm
api-consumption.test.cjs into per-area files (snp/cose/caci) sharing a
common support.cjs harness (package loader, fixtures, error assertions).
Run the whole directory in CI instead of the single file.
